requirements.txt

beautifulsoup4
newspaper3k
selenium
pandas
numpy
requests
yfinance
matplotlib
Pillow
google-generativeai
scikit-learn
tensorflow
selenium
tensorflow.keras
lxml[html_clean]
